+++ frontend/src/App.js
@@ -1,13 +1,23 @@
 import React from 'react';
 import './App.css';
-import FrontPage from './components/FrontPage';
+import { Routes, Route } from 'react-router-dom';
+import About from './components/About';
+import Store from './components/Store';
+import FrontPage from './components/frontpage/FrontPage';
 
 function App() {
   return (
-    <div className="App">
-      <h1>Sensoria</h1>
-      <FrontPage />
-    </div>
+    <>
+    <div className="container mb-5">
+        <Routes>
+          <Route exact path="/" element={<FrontPage/>} />
+          <Route path="/about" element={<About/>} />
+          <Route path="/store" element={<Store/>} />
+          <Route path="/client" element={<FrontPage/>} />
+          <Route path="/contact" element={<FrontPage/>} />
+          <Route path="/demo" element={<FrontPage/>} />
+        </Routes>
+      </div></>
   );
 }
