--- frontend/src/components/FrontPage.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function FrontPage(){
-    return(
-        <div>
-            <p>Welcome to the Front Page!</p>
-        </div>
-    )
-}
-
-export default FrontPage;
